refactor: generate raml docs using latest rmf-codegen, use proper description component for endpoints
