+++ packages/gatsby-theme-api-docs/src/components/description.js
