+++ websites/api-docs-smoke-test/src/api-specs/test/api.raml
@@ -15,6 +15,9 @@ annotationTypes:
     type: array
     items: string
     allowedTargets: TypeDeclaration
+  codeExamples:
+    type: object
+    allowedTargets: Method
   enumDescriptions:
      type: object
      (builtinType): object
