+++ websites/api-docs-smoke-test/src/api-specs/test/gen.properties
@@ -1,1 +1,1 @@
-hash=debd76840f836a2316255539a08342317964ec92
+hash=9c37368fd4abe306656e107ad1e40a4f5324e10e
