+++ websites/api-docs-smoke-test/src/api-specs/test/resources/ByProjectKeyResourceArtificiallyComplexPathUriParameterOneByUriParameterOneByUriParameterTwo.raml
@@ -89,3 +89,35 @@ post:
             default:
               strict: true
               value: !include ../examples/ByProjectKeyResourceArtificiallyComplexPathUriParameterOneByUriParameterOneByUriParameterTwoPost-202-default.json
+
+  (codeExamples):
+    curl: |-
+      curl -XPOST /{projectKey}/resource/artificially-complex/path/uri-parameter-one={uriParameterOne}/{uriParameterTwo} -i \
+      --header 'Authorization: Bearer YOUR_TOKEN \
+      --header 'Content-Type: application/json' \
+      --data-binary @- << EOF 
+      {
+        "eigthProperty" : "Should display 8th (1st in RAML)",
+        "createdBy" : {
+          "firstName" : "John",
+          "lastName" : "Doe"
+        },
+        "ninthProperty" : "Should display 9th (3rd in RAML)",
+        "key" : "Should display 3rd (4th in RAML)",
+        "tenthProperty" : "Should display 10th (5th in RAML)",
+        "custom" : "Should display 16th (6th in RAML)",
+        "eleventhPropery" : "Should display 11th (7th in RAML)",
+        "id" : "Should display 1st (8th in RAML)",
+        "twelthProperty" : "Should display 12th (9th in RAML)",
+        "lastModifiedBy" : {
+          "firstName" : "Jane",
+          "lastName" : "Doe"
+        },
+        "thirteenthProperty" : "Should display 13th (11th in RAML)",
+        "lastModifiedAt" : "2016-02-28T16:41:41.090Z",
+        "fourteenthProperty" : "Should display 14th (13th in RAML)",
+        "fifteenthProperty" : "Should display 15th (14th in definition)",
+        "createdAt" : "2015-02-28T16:41:41.090Z",
+        "version" : "Should display 2nd (16th in RAML)"
+      }
+      EOF
