+++ websites/api-docs-smoke-test/src/api-specs/test/resources/ByProjectKeyResourceSearchApiPattern.raml
@@ -87,3 +87,7 @@ get:
         default:
           strict: true
           value: "en"
+  (codeExamples):
+    curl: |-
+      curl -XGET /{projectKey}/resource/searchApiPattern -i \
+      --header 'Authorization: Bearer YOUR_TOKEN'
